openfaas-functions/image-processing-pillow/pyproject.toml

[tool.poetry]
name = "sentiment"
version = "0.1.0"
description = "Experiment facilities for the Fog"
authors = ["Volodia PAROL-GUARINO <volodia.parol-guarino@proton.me>"]
license = "MIT"

[tool.poetry.dependencies]
python = "~3.11"
# enoslib = { git = "https://gitlab.inria.fr/vparolgu/enoslib.git", rev = "f0c3c00a7665c2cff79b63b966f5a56b4b2ead18" }
flask = "*"
waitress = "*"
pillow = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

# Fixes <clock offset="localtime"/> to utc, leading to fix of k3s certificate being generated for an jour in the future from the pov of the vm
[tool.ruff]
line-length = 150

[tool.isort]
profile = "black"
